The connection configuration is a not a shared_ptr anymore
This was not a good idea to begin with. It implied that you could change
the connection configuration on the flight, which was not really the
case. The only impact would have been on the debug parameter.

Methods to actually update the debug parameter or the whole
configuration might be required on the long run.